+++ tools/idea-plugin/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- [Web Import] Add `Font Awesome` icons provider
+
 ## 1.4.0 - 2026-03-15
 
 ### Changed